requirements.txt
pandas
python-binance
tqdm
python-dotenv